resume form

+++ src/pages/api/resume.js
@@ -32,19 +32,22 @@ const handleGET = async (req, res) => {
     const { authorization } = req.headers
     try {
         const decoded = jwt.decode(authorization);
-        const resume = await Resume.find({ created_by: decoded?._id });
-        res.status(200)
-        res.send(resume)
+        if (decoded?._id) {
+            const resume = await Resume.find({ created_by: decoded?._id });
+            res.status(200)
+            res.send(resume)
+        } else {
+            throw new TypeError(res.status(401).send({ error: "Unauthorized" }));
+        }
     } catch (err) {
-        console.error('Error decoding token:', err);
+        res.status(400)
+        res.send(err);
     }
     // try {
     //     const resume = await Resume.find();
     //     res.status(200)
     //     res.send(resume)
     // } catch (err) {
-    //     // res.status(400)
-    //     res.send(err);
     // }
 };
 
@@ -53,10 +56,14 @@ const handlePOST = async (req, res) => {
 
     const decoded = jwt.decode(authorization);
     try {
-        const resume = new Resume({ ...req.body, ...{ created_by: decoded?._id } });
-        const createResume = await resume.save();
-        res.status(201)
-        res.send(createResume)
+        if (decoded?._id) {
+            const resume = new Resume({ ...req.body, ...{ created_by: decoded?._id } });
+            await resume.save();
+            res.status(201)
+            res.send({ data: { message: "Details Created Successfully" } })
+        } else {
+            throw new TypeError(res.status(401).send({ error: "Unauthorised" }));
+        }
     } catch (error) {
         console.log(error)
         res.status(400).send(error);
@@ -64,10 +71,15 @@ const handlePOST = async (req, res) => {
 };
 
 const handleDeleteById = async (req, res) => {
+    const { authorization } = req.headers
+
+    const decoded = jwt.decode(authorization);
     try {
-        const resume = await Resume.findByIdAndDelete(req.body)
-        res.status(201);
-        res.send(resume)
+        if (decoded?._id) {
+            await Resume.findByIdAndDelete(req.body)
+            res.status(201);
+            res.send({ data: { message: "Successfully deleted" } })
+        }
     } catch (err) {
         console.log(err)
         res.status(400).send(err);
